% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlySpectra.R
\name{plotlySpectra}
\alias{plotlySpectra}
\title{Interactive visualisation of a single spectrum}
\usage{
plotlySpectra(object)
}
\arguments{
\item{object}{A instance of class \code{Spectra} of length 1.}
}
\value{
A \code{plotly} object.
}
\description{
The \code{plotlySpectra()} function displays a single spectrum stored
in a \code{Spectra} object using the \code{plotly::plot_ly()} function.

See \code{?SpectraVis} for an example.
}
\examples{
library(msdata)
fl <- system.file("TripleTOF-SWATH", "PestMix1_DDA.mzML", package = "msdata")
pest_ms2 <- filterMsLevel(Spectra(fl), 2L)

plotlySpectra(pest_ms2[950])
}
\author{
Laurent Gatto, Johannes Rainer
}
