% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlySpectraMirror.R
\name{plotlySpectraMirror}
\alias{plotlySpectraMirror}
\title{Interactive visualisation of a mirror plot}
\usage{
plotlySpectraMirror(
  x,
  y,
  xLabel = "",
  xColor = "#737373",
  yLabel = "",
  yColor = "#737373",
  ppm = 20,
  tolerance = 0,
  matchColor = "#80B1D3"
)
}
\arguments{
\item{x}{A \code{\link[=Spectra]{Spectra()}} object of length 1.}

\item{y}{A \code{\link[=Spectra]{Spectra()}} object of length 1.}

\item{xLabel}{Optional \code{character(1)} with the label (name) of \code{x}.}

\item{xColor}{Color for peaks of spectrum \code{x}.}

\item{yLabel}{Optional \code{character(1)} with the label (name) of \code{y}.}

\item{yColor}{Color for peaks of spectrum \code{y}.}

\item{ppm}{\code{numeric(1)} with the m/z relative acceptable difference (in
ppm) for peaks to be considered matching (see \code{\link[=common]{common()}} for more
details).}

\item{tolerance}{\code{numeric(1)} with the absolute acceptable difference of
m/z values for peaks to be considered matching (see \code{\link[=common]{common()}} for more
details).}

\item{matchColor}{Color for matching peaks.}
}
\value{
A \code{plotly} object
}
\description{
The \code{plotlySpectraMirror} function creates an interactive \emph{mirror plot}
comparing two spectra \code{x} and \code{y} with each other. See also
\code{\link[=plotSpectraMirror]{plotSpectraMirror()}} for a non-interactive version of this plot.
}
\examples{

## Load example data.
library(msdata)
fl <- system.file("TripleTOF-SWATH", "PestMix1_DDA.mzML", package = "msdata")
pest_ms2 <- filterMsLevel(Spectra(fl), 2L)

plotlySpectraMirror(pest_ms2[948], pest_ms2[950])

plotlySpectraMirror(pest_ms2[948], pest_ms2[959], xLabel = "query",
    yLabel = "target")
}
\author{
Johannes Rainer
}
