% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraVis.R
\docType{package}
\name{SpectraVis}
\alias{SpectraVis-package}
\alias{SpectraVis}
\title{SpectraVis: Visualising and Exploring Spectra Data}
\description{
This package defines a set of helper function to visualise and
explore mass spectrometry data stored as Spectra objects.
}
\section{SpectraVis functions}{

\itemize{
\item plotlySpectra: Interactive visualisation of a single spectrum.
\item browseSpectra: Browse spectra in a Spectra object.
}
}

\examples{

library(MsDataHub)
f <- MS3TMT10_01022016_32917.33481.mzML.gz()
sp <- Spectra(f)
sp

if (interactive())
   browseSpectra(sp)

## Use Ctrl+C to interrupt R and stop the application

if (interactive())
   plotlySpectra(sp[1])
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RforMassSpectrometry/SpectraVis}
  \item Report bugs at \url{https://github.com/RforMassSpectrometry/SpectraVis/issues}
}

}
\author{
\strong{Maintainer}: Laurent Gatto \email{laurent.gatto@uclouvain.be} (\href{https://orcid.org/0000-0002-1520-2268}{ORCID})

Authors:
\itemize{
  \item Laurent Gatto \email{laurent.gatto@uclouvain.be} (\href{https://orcid.org/0000-0002-1520-2268}{ORCID})
  \item Johannes Rainer \email{Johannes.Rainer@eurac.edu} (\href{https://orcid.org/0000-0002-6977-7147}{ORCID})
}

}
