% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseSpectra.R
\name{browseSpectra}
\alias{browseSpectra}
\title{Browse spectra in a Spectra object}
\usage{
browseSpectra(object)
}
\arguments{
\item{object}{A non-empty instance of class \code{Spectra}.}
}
\value{
An object that represents the app.
}
\description{
The \code{browseSpectra()} function opens a simple shiny application
that allows to browse trough the individual scans of a \code{Spectra}
object. The title of the scans contains the MS level, the
retention time, if available the precursor m/z and precursor
charge and a peptides sequence (as defined in the \code{sequence}
spectra variable) and the index of the scan. If a scans has a
sequence (as defined in the \code{sequence} spectra variable), the
matching fragments (as defined by \code{PSMatch::calculateFragments()} are
labelled, otherwise the most intense peaks (as defined as those
that have an intensity as high as half the highest peak) are
labelled with their m/z.

See \code{?SpectraVis} for an example.
}
\author{
Laurent Gatto
}
