% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlySpectra.R
\name{plotlySpectra}
\alias{plotlySpectra}
\title{Interactive visualisation of a single spectrum}
\usage{
plotlySpectra(object)
}
\arguments{
\item{object}{A instance of class \code{Spectra} of length 1.}
}
\value{
A \code{plotly} object.
}
\description{
The \code{plotlySpectrum()} function displays a single spectrum stored
in a \code{Spectra} object using the \code{plotly::plot_ly()} function.

See \code{?SpectraVis} for an example.
}
\author{
Laurent Gatto
}
