% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraVis.R
\docType{package}
\name{SpectraVis}
\alias{SpectraVis}
\title{SpectraVis: Visualising and Exploring Spectra Data}
\description{
This package defines a set of helper function to visualise and
explore mass spectrometry data stored as Spectra objects.
}
\section{SpectraVis functions}{

\itemize{
\item plotlySpectra: Interactive visualisation of a single spectrum.
\item browseSpectra: Browse spectra in a Spectra object.
}
}

\examples{

f <- msdata::proteomics(pattern = "MS3TMT10", full.names = TRUE)
sp <- Spectra(f)
sp

if (interactive())
   browseSpectra(sp)

## Use Ctrl+C to interrupt R and stop the application

if (interactive())
   plotlySpectra(sp[1])
}
